requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
requests>=2.28.0
urllib3>=2.0.0  # For TLS 1.3 support

# Database
pymongo>=4.3.0
dnspython>=2.3.0  # Required for MongoDB SRV connection strings

# Pydantic for settings and validation
pydantic>=1.10.0
pydantic-settings>=2.0.0

# Rich for console output
rich>=12.0.0

# For image processing
pillow>=11.2.1
opencv-python-headless>=4.8.0
numpy>=1.24.0

# AWS Services (SQS + S3/Wasabi)
boto3>=1.28.0
botocore>=1.31.0

# Async Support (for production async workflows)
asyncio>=3.4.3
aiofiles>=23.0.0  # For async file operations